Revert "🐛(front) optimize chat"
This reverts commit 69bf2cab5d67447019a37db030bc44487f36b842.
