+++ src/frontend/apps/conversations/src/features/chat/components/Chat.tsx
@@ -1,26 +1,45 @@
-import { Message, SourceUIPart } from '@ai-sdk/ui-utils';
+import {
+  Message,
+  ReasoningUIPart,
+  SourceUIPart,
+  ToolInvocationUIPart,
+} from '@ai-sdk/ui-utils';
 import { Modal, ModalSize } from '@openfun/cunningham-react';
+import 'katex/dist/katex.min.css'; // `rehype-katex` does not import the CSS for you
 import { useRouter } from 'next/router';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import type { ChangeEvent, FormEvent } from 'react';
 import { useTranslation } from 'react-i18next';
+import { MarkdownHooks } from 'react-markdown';
+import rehypeKatex from 'rehype-katex';
+import rehypePrettyCode from 'rehype-pretty-code';
+import remarkGfm from 'remark-gfm';
+import remarkMath from 'remark-math';
 
 import { APIError, errorCauses, fetchAPI } from '@/api';
-import { Box, Loader, Text } from '@/components';
+import { Box, Icon, Loader, Text } from '@/components';
 import { useUploadFile } from '@/features/attachments/hooks/useUploadFile';
 import { useChat } from '@/features/chat/api/useChat';
 import { getConversation } from '@/features/chat/api/useConversation';
 import { useCreateChatConversation } from '@/features/chat/api/useCreateConversation';
+import {
+  LLMModel,
+  useLLMConfiguration,
+} from '@/features/chat/api/useLLMConfiguration';
+import { AttachmentList } from '@/features/chat/components/AttachmentList';
 import { ChatError } from '@/features/chat/components/ChatError';
+import { CodeBlock } from '@/features/chat/components/CodeBlock';
+import { FeedbackButtons } from '@/features/chat/components/FeedbackButtons';
 import { InputChat } from '@/features/chat/components/InputChat';
-import { useChatPreferencesStore } from '@/features/chat/stores/useChatPreferencesStore';
-import { usePendingChatStore } from '@/features/chat/stores/usePendingChatStore';
-import { useScrollStore } from '@/features/chat/stores/useScrollStore';
+import { SourceItemList } from '@/features/chat/components/SourceItemList';
+import { ToolInvocationItem } from '@/features/chat/components/ToolInvocationItem';
+import { useClipboard } from '@/hook';
 import { useResponsiveStore } from '@/stores';
 
-import { useModelSelection, useSourceMetadataCache } from '../hooks';
-
-import { ChatMessage } from './ChatMessage';
+import { useSourceMetadataCache } from '../hooks';
+import { useChatPreferencesStore } from '../stores/useChatPreferencesStore';
+import { usePendingChatStore } from '../stores/usePendingChatStore';
+import { useScrollStore } from '../stores/useScrollStore';
 
 // Define Attachment type locally (mirroring backend structure)
 export interface Attachment {
@@ -35,16 +54,22 @@ export const Chat = ({
   initialConversationId: string | undefined;
 }) => {
   const { t } = useTranslation();
+  const copyToClipboard = useClipboard();
   const { isMobile } = useResponsiveStore();
+
   const streamProtocol = 'data'; // or 'text'
 
-  const { forceWebSearch, toggleForceWebSearch } = useChatPreferencesStore();
+  const {
+    forceWebSearch,
+    toggleForceWebSearch,
+    selectedModelHrid,
+    setSelectedModelHrid,
+  } = useChatPreferencesStore();
 
-  const { selectedModel, handleModelSelect } = useModelSelection();
+  const { data: llmConfig } = useLLMConfiguration();
+  const [selectedModel, setSelectedModel] = useState<LLMModel | null>(null);
 
-  // Use custom hook for conversation sync - we'll update it after useChat
   const [conversationId, setConversationId] = useState(initialConversationId);
-
   const apiUrl = conversationId
     ? `chats/${conversationId}/conversation/?protocol=${streamProtocol}`
     : `chats/conversation/?protocol=${streamProtocol}`;
@@ -58,6 +83,36 @@ export const Chat = ({
     (window as { globalForceWebSearch?: boolean }).globalForceWebSearch =
       forceWebSearch;
   }, [forceWebSearch]);
+
+  // Update selected model when LLM config loads
+  useEffect(() => {
+    if (llmConfig?.models && !selectedModel) {
+      let modelToSelect: LLMModel | undefined;
+
+      if (selectedModelHrid) {
+        // Try to find the previously selected model
+        modelToSelect = llmConfig.models.find(
+          (model) =>
+            model.hrid === selectedModelHrid && model.is_active !== false,
+        );
+      }
+
+      // If no saved model or saved model not found/inactive, use default
+      if (!modelToSelect) {
+        modelToSelect = llmConfig.models.find((model) => model.is_default);
+      }
+
+      if (modelToSelect) {
+        setSelectedModel(modelToSelect);
+        setSelectedModelHrid(modelToSelect.hrid);
+      }
+    }
+  }, [llmConfig, selectedModel, selectedModelHrid, setSelectedModelHrid]);
+
+  const handleModelSelect = (model: LLMModel) => {
+    setSelectedModel(model);
+    setSelectedModelHrid(model.hrid);
+  };
 
   const router = useRouter();
   const [files, setFiles] = useState<FileList | null>(null);
@@ -95,22 +150,16 @@ export const Chat = ({
 
   const [initialConversationMessages, setInitialConversationMessages] =
     useState<Message[] | undefined>(undefined);
-  const {
-    input: pendingInput,
-    files: _pendingFiles,
-    setPendingChat,
-    clearPendingChat: _clearPendingChat,
-  } = usePendingChatStore();
-  const [hasInitialized, setHasInitialized] = useState(false);
-
   const [pendingFirstMessage, setPendingFirstMessage] = useState<{
     event: FormEvent<HTMLFormElement>;
     attachments?: Attachment[];
     forceWebSearch?: boolean;
   } | null>(null);
+  const [shouldAutoSubmit, setShouldAutoSubmit] = useState(false);
   const [shouldRetry, setShouldRetry] = useState(false);
   const retryOriginalInputRef = useRef<string>('');
   const retryOriginalFilesRef = useRef<FileList | null>(null);
+  const [hasInitialized, setHasInitialized] = useState(false);
   const [streamingMessageHeight, setStreamingMessageHeight] = useState<
     number | null
   >(null);
@@ -125,6 +174,23 @@ export const Chat = ({
 
   const { mutate: createChatConversation } = useCreateChatConversation();
 
+  // Zustand store for pending chat state
+  const {
+    input: pendingInput,
+    files: pendingFiles,
+    setPendingChat,
+    clearPendingChat,
+  } = usePendingChatStore();
+
+  const scrollToBottom = useCallback(() => {
+    if (chatContainerRef.current) {
+      chatContainerRef.current.scrollTo({
+        top: chatContainerRef.current.scrollHeight,
+        behavior: hasInitialized ? 'smooth' : 'auto',
+      });
+    }
+  }, [hasInitialized]);
+
   // Show error modal for upload errors
   useEffect(() => {
     if (isErrorAttachment && errorAttachment) {
@@ -136,18 +202,15 @@ export const Chat = ({
   }, [isErrorAttachment, errorAttachment, t]);
 
   // Handle errors from the chat API
-  const onErrorChat = useCallback(
-    (error: Error) => {
-      if (error.message === 'attachment_summary_not_supported') {
-        setChatErrorModal({
-          title: t('Attachment summary not supported'),
-          message: t('The summary feature is not supported yet.'),
-        });
-      }
-      console.error('Chat error:', error);
-    },
-    [t],
-  );
+  const onErrorChat = (error: Error) => {
+    if (error.message === 'attachment_summary_not_supported') {
+      setChatErrorModal({
+        title: t('Attachment summary not supported'),
+        message: t('The summary feature is not supported yet.'),
+      });
+    }
+    console.error('Chat error:', error);
+  };
 
   const {
     messages,
@@ -159,20 +222,64 @@ export const Chat = ({
     setMessages,
   } = useChat({
     id: conversationId,
-    // Ne pas réinitialiser les messages si on est en train de créer une conversation
-    initialMessages: pendingFirstMessage
-      ? undefined
-      : initialConversationMessages,
+    initialMessages: initialConversationMessages,
     api: apiUrl,
     streamProtocol: streamProtocol,
     sendExtraMessageFields: true,
     onError: onErrorChat,
   });
 
-  // Ref pour messages pour éviter de recréer openSources à chaque changement
-  const messagesRef = useRef(messages);
-  messagesRef.current = messages;
+  const stopGeneration = async () => {
+    stopChat();
 
+    const response = await fetchAPI(`chats/${conversationId}/stop-streaming/`, {
+      method: 'POST',
+    });
+
+    if (!response.ok) {
+      throw new APIError(
+        'Failed to stop the conversation',
+        await errorCauses(response),
+      );
+    }
+  };
+
+  const toggleWebSearch = () => {
+    toggleForceWebSearch();
+  };
+
+  const handleStop = () => {
+    void stopGeneration();
+  };
+
+  const handleSubmitWrapper = (event: FormEvent<HTMLFormElement>) => {
+    void handleSubmit(event);
+  };
+
+  const handleRetry = () => {
+    if (!lastSubmissionRef.current || !setMessages) {
+      return;
+    }
+
+    const { input: lastInput, files: lastFiles } = lastSubmissionRef.current;
+
+    const lastAssistantIndex = messages.findLastIndex(
+      (msg) => msg.role === 'assistant',
+    );
+    if (lastAssistantIndex !== -1) {
+      setMessages(messages.filter((_, index) => index !== lastAssistantIndex));
+    }
+
+    retryOriginalInputRef.current = input;
+    retryOriginalFilesRef.current = files;
+    handleInputChange({
+      target: { value: lastInput },
+    } as ChangeEvent<HTMLTextAreaElement>);
+    setFiles(lastFiles);
+    setShouldRetry(true);
+  };
+
+  // Précharger les métadonnées des sources dès que les messages arrivent
   useEffect(() => {
     messages.forEach((message) => {
       if (message.parts) {
@@ -187,265 +294,28 @@ export const Chat = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [messages]);
 
-  // Custom handleSubmit to include attachments and handle chat creation
-  const handleSubmit = useCallback(
-    async (event: FormEvent<HTMLFormElement>) => {
-      event.preventDefault();
-
-      // Upload files to server and get URLs
-      let attachments: Attachment[] = [];
-      if (files && files.length > 0 && conversationId) {
-        try {
-          setIsUploadingFiles(true);
-          const uploadPromises = Array.from(files).map(async (file) => {
-            const url = await uploadFile(file);
-
-            return {
-              name: file.name,
-              contentType: file.type,
-              url: url,
-            };
-          });
-          attachments = await Promise.all(uploadPromises);
-          setIsUploadingFiles(false);
-        } catch (error) {
-          setIsUploadingFiles(false);
-          console.error('File upload error:', error);
-          setChatErrorModal({
-            title: t('Upload Error'),
-            message: t('Failed to upload files. Please try again.'),
-          });
-          return;
-        }
-      }
-
-      if (!conversationId) {
-        // Save the event and files, then create the chat
-        setPendingFirstMessage({ event, attachments, forceWebSearch });
-        // Save input and files to Zustand store before navigation
-        setPendingChat(input, files);
-        void createChatConversation(
-          { title: input.length > 100 ? `${input.slice(0, 97)}...` : input },
-          {
-            onSuccess: (data: { id: string }) => {
-              setConversationId(data.id);
-              // Update the URL to /chat/[id]/
-              void router.push(`/chat/${data.id}/`);
-              // Le message sera envoyé via le useEffect qui attend que useChat soit prêt
-            },
-          },
-        );
-        return;
-      }
-
-      // Prepare options with attachments
-      const options: Record<string, unknown> = {};
-      if (attachments.length > 0) {
-        options.experimental_attachments = attachments;
-      }
-
-      lastSubmissionRef.current = {
-        input,
-        files,
-        event,
-        options: Object.keys(options).length > 0 ? options : undefined,
-      };
-
-      if (Object.keys(options).length > 0) {
-        baseHandleSubmit(event, options);
-      } else {
-        baseHandleSubmit(event);
-      }
-      // Attendre un peu avant de vider les fichiers pour s'assurer qu'ils sont traités
-      setTimeout(() => {
-        setFiles(null);
-        if (fileInputRef.current) {
-          fileInputRef.current.value = '';
-        }
-      }, 100);
-    },
-    [
-      files,
-      conversationId,
-      uploadFile,
-      t,
-      forceWebSearch,
-      setPendingChat,
-      input,
-      createChatConversation,
-      setConversationId,
-      router,
-      baseHandleSubmit,
-    ],
-  );
-
-  // Synchronize conversationId state with prop when it changes
-  useEffect(() => {
-    // Ne réinitialiser que si on change vraiment de conversation (pas lors de la création)
-    if (
-      initialConversationId &&
-      initialConversationId !== conversationId &&
-      !pendingFirstMessage
-    ) {
-      setConversationId(initialConversationId);
-      // Reset input when conversation changes
-      handleInputChange({
-        target: { value: '' },
-      } as ChangeEvent<HTMLTextAreaElement>);
-      setHasInitialized(false);
-    } else if (
-      !initialConversationId &&
-      conversationId &&
-      !pendingFirstMessage
-    ) {
-      // Si on n'a plus d'initialConversationId mais qu'on a un conversationId, ne rien faire
-      // (on est peut-être en train de créer une conversation)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [initialConversationId, conversationId, pendingFirstMessage]);
-
-  useEffect(() => {
-    if (
-      conversationId &&
-      pendingFirstMessage &&
-      status === 'ready' &&
-      apiUrl.includes(conversationId)
-    ) {
-      const pending = pendingFirstMessage;
-
-      // Préparer les options avec les attachments
-      const options: Record<string, unknown> = {};
-      if (pending.attachments && pending.attachments.length > 0) {
-        options.experimental_attachments = pending.attachments;
-      }
-
-      // Mettre à jour lastSubmissionRef pour le retry
-      lastSubmissionRef.current = {
-        input,
-        files,
-        event: pending.event,
-        options: Object.keys(options).length > 0 ? options : undefined,
-      };
-
-      setPendingFirstMessage(null);
-      requestAnimationFrame(() => {
-        requestAnimationFrame(() => {
-          if (Object.keys(options).length > 0) {
-            baseHandleSubmit(pending.event, options);
-          } else {
-            baseHandleSubmit(pending.event);
-          }
-
-          // Nettoyer les fichiers après l'envoi
-          setFiles(null);
-          if (fileInputRef.current) {
-            fileInputRef.current.value = '';
-          }
-        });
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [conversationId, status, baseHandleSubmit, pendingFirstMessage, apiUrl]);
-
-  // Fetch initial conversation messages if initialConversationId is provided
-  useEffect(() => {
-    hasScrolledToBottomOnLoadRef.current = false;
-    let ignore = false;
-    async function fetchInitialMessages() {
-      if (initialConversationId && !pendingInput && !pendingFirstMessage) {
-        try {
-          const conversation = await getConversation({
-            id: initialConversationId,
-          });
-          if (!ignore) {
-            setInitialConversationMessages(conversation.messages);
-            setHasInitialized(true);
-          }
-        } catch {
-          if (!ignore) {
-            setInitialConversationMessages([]);
-            setHasInitialized(true);
-          }
-        }
-      }
-    }
-    void fetchInitialMessages();
-    return () => {
-      ignore = true;
-    };
-  }, [initialConversationId, pendingInput, pendingFirstMessage]);
-
-  const scrollToBottom = useCallback(() => {
-    if (chatContainerRef.current) {
-      chatContainerRef.current.scrollTo({
-        top: chatContainerRef.current.scrollHeight,
-        behavior: hasInitialized ? 'smooth' : 'auto',
-      });
-    }
-  }, [hasInitialized]);
-
-  const stopGeneration = useCallback(async () => {
-    stopChat();
-
-    if (!conversationId) {
+  const openSources = (messageId: string) => {
+    if (isSourceOpen === messageId) {
+      setIsSourceOpen(null);
       return;
     }
-
-    const response = await fetchAPI(`chats/${conversationId}/stop-streaming/`, {
-      method: 'POST',
-    });
-
-    if (!response.ok) {
-      throw new APIError(
-        'Failed to stop the conversation',
-        await errorCauses(response),
+    const message = messages.find((msg) => msg.id === messageId);
+    if (message?.parts) {
+      const sourceParts = message.parts.filter(
+        (part): part is SourceUIPart => part.type === 'source',
       );
+      if (sourceParts.length > 0) {
+        setIsSourceOpen(messageId);
+      }
     }
-  }, [stopChat, conversationId]);
-
-  const toggleWebSearch = useCallback(() => {
-    toggleForceWebSearch();
-  }, [toggleForceWebSearch]);
-
-  const handleStop = useCallback(() => {
-    void stopGeneration();
-  }, [stopGeneration]);
-
-  const handleSubmitWrapper = useCallback(
-    (event: FormEvent<HTMLFormElement>) => {
-      void handleSubmit(event);
-    },
-    [handleSubmit],
-  );
-
-  // Utiliser un ref pour messages pour éviter de recréer la fonction à chaque changement
-  const openSources = useCallback(
-    (messageId: string) => {
-      if (isSourceOpen === messageId) {
-        setIsSourceOpen(null);
-        return;
-      }
-      const message = messagesRef.current.find((msg) => msg.id === messageId);
-      if (message?.parts) {
-        const sourceParts = message.parts.filter(
-          (part) => part.type === 'source',
-        );
-        if (sourceParts.length > 0) {
-          setIsSourceOpen(messageId);
-        }
-      }
-    },
-    [isSourceOpen], // Plus besoin de messages dans les dépendances
-  );
-
-  // Mémoriser le calcul du dernier index assistant (évite de le recalculer pour chaque message)
-  const lastAssistantIndex = useMemo(
-    () => messages.findLastIndex((msg) => msg.role === 'assistant'),
-    [messages],
-  );
+  };
 
   // Calculer la hauteur pour le message de streaming
   const calculateStreamingHeight = useCallback(() => {
+    if (messages.length <= 2) {
+      return;
+    }
+
     if (chatContainerRef.current) {
       const container = chatContainerRef.current;
       const containerHeight = container.clientHeight;
@@ -476,6 +346,7 @@ export const Chat = ({
     }
   }, [messages]);
 
+  // Détecter l'arrivée d'un nouveau message user et retirer la hauteur de l'ancien
   useEffect(() => {
     if (status === 'streaming') {
       return;
@@ -499,7 +370,8 @@ export const Chat = ({
     if (status === 'submitted' || status === 'streaming') {
       calculateStreamingHeight();
     }
-  }, [status, calculateStreamingHeight]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [status]);
 
   // Scroller vers la question au moment du submit
   useEffect(() => {
@@ -519,7 +391,108 @@ export const Chat = ({
 
       messageElement?.scrollIntoView({ block: 'start', behavior: 'smooth' });
     });
-  }, [status, messages]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [status]);
+
+  // Synchronize conversationId state with prop when it changes (e.g., after navigation)
+  useEffect(() => {
+    setConversationId(initialConversationId);
+    // Reset input when conversation changes
+    if (initialConversationId !== conversationId) {
+      handleInputChange({
+        target: { value: '' },
+      } as ChangeEvent<HTMLTextAreaElement>);
+      setHasInitialized(false); // Réinitialiser pour permettre le scroll au prochain chargement
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [initialConversationId, conversationId]);
+
+  // On mount, if there is pending input/files, initialize state and set flag
+  useEffect(() => {
+    if (
+      (pendingInput && pendingInput.trim()) ||
+      (pendingFiles && pendingFiles.length > 0)
+    ) {
+      if (pendingInput) {
+        const syntheticEvent = {
+          target: { value: pendingInput },
+        } as ChangeEvent<HTMLInputElement>;
+        handleInputChange(syntheticEvent);
+      }
+      if (pendingFiles) {
+        setFiles(pendingFiles);
+      }
+      setShouldAutoSubmit(true);
+      clearPendingChat();
+    } else {
+      clearPendingChat();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  // When shouldAutoSubmit is set, and input/files are ready, submit
+  useEffect(() => {
+    if (shouldAutoSubmit && (input.trim() || (files && files.length > 0))) {
+      // Create a synthetic event for form submission
+      const form = document.createElement('form');
+      const syntheticFormEvent = {
+        preventDefault: () => {},
+        target: form,
+      } as unknown as FormEvent<HTMLFormElement>;
+      void handleSubmit(syntheticFormEvent);
+      setShouldAutoSubmit(false);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [shouldAutoSubmit, input, files]);
+
+  useEffect(() => {
+    if (
+      shouldRetry &&
+      lastSubmissionRef.current &&
+      input === lastSubmissionRef.current.input
+    ) {
+      const { event } = lastSubmissionRef.current;
+
+      void handleSubmit(event);
+      handleInputChange({
+        target: { value: retryOriginalInputRef.current },
+      } as ChangeEvent<HTMLTextAreaElement>);
+      setFiles(retryOriginalFilesRef.current);
+
+      setShouldRetry(false);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [shouldRetry, input, files]);
+
+  // Fetch initial conversation messages if initialConversationId is provided and no pending input
+  useEffect(() => {
+    hasScrolledToBottomOnLoadRef.current = false; // Réinitialiser au début du chargement
+    let ignore = false;
+    async function fetchInitialMessages() {
+      if (initialConversationId && !pendingInput) {
+        try {
+          const conversation = await getConversation({
+            id: initialConversationId,
+          });
+          if (!ignore) {
+            setInitialConversationMessages(conversation.messages);
+            setHasInitialized(true);
+          }
+        } catch {
+          // Optionally handle error (e.g., setInitialConversationMessages([]) or show error)
+          if (!ignore) {
+            setInitialConversationMessages([]);
+            setHasInitialized(true);
+          }
+        }
+      }
+    }
+    void fetchInitialMessages();
+    return () => {
+      ignore = true;
+    };
+    // Only run when initialConversationId or pendingInput changes
+  }, [initialConversationId, pendingInput]);
 
   useEffect(() => {
     if (
@@ -540,56 +513,109 @@ export const Chat = ({
         });
       });
     }
-  }, [hasInitialized, conversationId, messages.length]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [hasInitialized, messages.length]);
 
-  const handleRetry = useCallback(() => {
-    if (!lastSubmissionRef.current || !setMessages) {
+  // Custom handleSubmit to include attachments and handle chat creation
+  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+
+    // Upload files to server and get URLs
+    let attachments: Attachment[] = [];
+    if (files && files.length > 0 && conversationId) {
+      try {
+        setIsUploadingFiles(true);
+        const uploadPromises = Array.from(files).map(async (file) => {
+          const url = await uploadFile(file);
+
+          return {
+            name: file.name,
+            contentType: file.type,
+            url: url,
+          };
+        });
+        attachments = await Promise.all(uploadPromises);
+        setIsUploadingFiles(false);
+      } catch (error) {
+        setIsUploadingFiles(false);
+        console.error('File upload error:', error);
+        setChatErrorModal({
+          title: t('Upload Error'),
+          message: t('Failed to upload files. Please try again.'),
+        });
+        return;
+      }
+    }
+
+    if (!conversationId) {
+      // Save the event and files, then create the chat
+      setPendingFirstMessage({ event, attachments, forceWebSearch });
+      // Save input and files to Zustand store before navigation
+      setPendingChat(input, files);
+      void createChatConversation(
+        { title: input.length > 100 ? `${input.slice(0, 97)}...` : input },
+        {
+          onSuccess: (data) => {
+            setConversationId(data.id);
+            // Update the URL to /chat/[id]/
+            void router.push(`/chat/${data.id}/`);
+            // After setting the conversationId, submit the pending message
+            setTimeout(() => {
+              if (pendingFirstMessage) {
+                // Prepare options with attachments
+                const options: Record<string, unknown> = {};
+                if (
+                  pendingFirstMessage.attachments &&
+                  pendingFirstMessage.attachments.length > 0
+                ) {
+                  options.experimental_attachments =
+                    pendingFirstMessage.attachments;
+                }
+
+                if (Object.keys(options).length > 0) {
+                  baseHandleSubmit(pendingFirstMessage.event, options);
+                } else {
+                  baseHandleSubmit(pendingFirstMessage.event);
+                }
+                setFiles(null);
+                if (fileInputRef.current) {
+                  fileInputRef.current.value = '';
+                }
+                setPendingFirstMessage(null);
+              }
+            }, 0);
+          },
+        },
+      );
       return;
     }
 
-    const { input: lastInput, files: lastFiles } = lastSubmissionRef.current;
-
-    const lastAssistantIndex = messages.findLastIndex(
-      (msg) => msg.role === 'assistant',
-    );
-    if (lastAssistantIndex !== -1) {
-      setMessages(messages.filter((_, index) => index !== lastAssistantIndex));
+    // Prepare options with attachments
+    const options: Record<string, unknown> = {};
+    if (attachments.length > 0) {
+      options.experimental_attachments = attachments;
     }
 
-    retryOriginalInputRef.current = input;
-    retryOriginalFilesRef.current = files;
-    handleInputChange({
-      target: { value: lastInput },
-    } as ChangeEvent<HTMLTextAreaElement>);
-    setFiles(lastFiles);
-    setShouldRetry(true);
-  }, [
-    lastSubmissionRef,
-    setMessages,
-    messages,
-    input,
-    files,
-    handleInputChange,
-  ]);
+    lastSubmissionRef.current = {
+      input,
+      files,
+      event,
+      options: Object.keys(options).length > 0 ? options : undefined,
+    };
 
-  useEffect(() => {
-    if (
-      shouldRetry &&
-      lastSubmissionRef.current &&
-      input === lastSubmissionRef.current.input
-    ) {
-      const { event } = lastSubmissionRef.current;
-
-      void handleSubmit(event);
-      handleInputChange({
-        target: { value: retryOriginalInputRef.current },
-      } as ChangeEvent<HTMLTextAreaElement>);
-      setFiles(retryOriginalFilesRef.current);
-
-      setShouldRetry(false);
+    if (Object.keys(options).length > 0) {
+      baseHandleSubmit(event, options);
+    } else {
+      baseHandleSubmit(event);
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [shouldRetry, input, files]);
+    // Attendre un peu avant de vider les fichiers pour s'assurer qu'ils sont traités
+    setTimeout(() => {
+      setFiles(null);
+      if (fileInputRef.current) {
+        fileInputRef.current.value = '';
+      }
+    }, 100);
+  };
 
   return (
     <Box
@@ -623,30 +649,333 @@ export const Chat = ({
             {messages.map((message, index) => {
               const isLastMessage = index === messages.length - 1;
               const isLastAssistantMessageInConversation =
-                message.role === 'assistant' && index === lastAssistantIndex;
-              const shouldApplyStreamingHeight = Boolean(
-                isLastAssistantMessageInConversation && isLastMessage,
-              );
+                message.role === 'assistant' &&
+                index ===
+                  messages.findLastIndex((msg) => msg.role === 'assistant');
+              const isFirstConversationMessage = messages.length <= 2;
+              const shouldApplyStreamingHeight =
+                isLastAssistantMessageInConversation &&
+                isLastMessage &&
+                streamingMessageHeight &&
+                !isFirstConversationMessage;
               const isCurrentlyStreaming =
                 isLastAssistantMessageInConversation &&
                 (status === 'streaming' || status === 'submitted');
 
               return (
-                <ChatMessage
+                <Box
                   key={message.id}
-                  message={message}
-                  isLastAssistantMessageInConversation={
-                    isLastAssistantMessageInConversation
-                  }
-                  shouldApplyStreamingHeight={shouldApplyStreamingHeight}
-                  streamingMessageHeight={streamingMessageHeight}
-                  isCurrentlyStreaming={isCurrentlyStreaming}
-                  status={isCurrentlyStreaming ? status : 'ready'} // Ne passer le vrai status que si nécessaire
-                  isSourceOpen={isSourceOpen}
-                  conversationId={conversationId}
-                  onOpenSources={openSources}
-                  getMetadata={getMetadata}
-                />
+                  data-message-id={message.id}
+                  $css={`
+                    display: flex;
+                    width: 100%;
+                    margin: auto;
+                    margin-bottom: ${isLastAssistantMessageInConversation ? '30px' : '0px'};
+                    color: var(--c--theme--colors--greyscale-850);
+                    padding-left: 12px;
+                    padding-right: 12px;
+                    max-width: 750px;
+                    text-align: left;
+                    overflow-wrap: anywhere;
+                    flex-direction: ${message.role === 'user' ? 'row-reverse' : 'row'};
+                  `}
+                >
+                  <Box
+                    $display="block"
+                    $width={`${message.role === 'user' ? 'auto' : '100%'}`}
+                  >
+                    {message.experimental_attachments &&
+                      message.experimental_attachments.length > 0 && (
+                        <Box>
+                          <AttachmentList
+                            attachments={message.experimental_attachments}
+                            isReadOnly={true}
+                          />
+                        </Box>
+                      )}
+                    <Box
+                      $radius="8px"
+                      $width={`${message.role === 'user' ? 'auto' : '100%'}`}
+                      $maxWidth="100%"
+                      $padding={`${message.role === 'user' ? '12px' : '0'}`}
+                      $margin={{ vertical: 'base' }}
+                      $background={`${message.role === 'user' ? '#EEF1F4' : 'white'}`}
+                      $css={`
+                      display: inline-block;
+                      float: right;
+                      ${shouldApplyStreamingHeight ? `min-height: ${streamingMessageHeight}px;` : ''}`}
+                    >
+                      {/* Message content */}
+                      {message.content && (
+                        <Box
+                          className="mainContent-chat"
+                          data-testid={
+                            message.role === 'assistant'
+                              ? 'assistant-message-content'
+                              : undefined
+                          }
+                          $padding={{ all: 'xxs' }}
+                        >
+                          <p className="sr-only">
+                            {message.role === 'user'
+                              ? t('You said: ')
+                              : t('Assistant IA replied: ')}
+                          </p>
+                          {message.role === 'user' ? (
+                            <Text
+                              as="p"
+                              $css="white-space: pre-wrap; display: block;"
+                              $theme="greyscale"
+                              $variation="850"
+                            >
+                              {message.content}
+                            </Text>
+                          ) : (
+                            <MarkdownHooks
+                              remarkPlugins={[remarkGfm, remarkMath]}
+                              rehypePlugins={[
+                                [
+                                  rehypePrettyCode,
+                                  {
+                                    theme: 'github-dark-dimmed',
+                                  },
+                                ],
+                                rehypeKatex,
+                              ]}
+                              components={{
+                                // Custom components for Markdown rendering
+                                // eslint-disable-next-line @typescript-eslint/no-unused-vars
+                                p: ({ node, ...props }) => (
+                                  <Text
+                                    as="p"
+                                    $css="display: block"
+                                    $theme="greyscale"
+                                    $variation="850"
+                                    {...props}
+                                  />
+                                ),
+                                a: ({ children, ...props }) => (
+                                  <a target="_blank" {...props}>
+                                    {children}
+                                  </a>
+                                ),
+                                // eslint-disable-next-line @typescript-eslint/no-unused-vars
+                                pre: ({ node, children, ...props }) => (
+                                  <CodeBlock {...props}>{children}</CodeBlock>
+                                ),
+                              }}
+                            >
+                              {message.content}
+                            </MarkdownHooks>
+                          )}
+                        </Box>
+                      )}
+
+                      <Box $direction="column" $gap="2">
+                        {isCurrentlyStreaming &&
+                          isLastAssistantMessageInConversation &&
+                          status === 'streaming' &&
+                          message.parts?.some(
+                            (part) =>
+                              part.type === 'tool-invocation' &&
+                              part.toolInvocation.toolName !==
+                                'document_parsing',
+                          ) && (
+                            <Box
+                              $direction="row"
+                              $align="center"
+                              $gap="6px"
+                              $width="100%"
+                              $maxWidth="750px"
+                              $margin={{
+                                all: 'auto',
+                                top: 'base',
+                                bottom: 'md',
+                              }}
+                            >
+                              <Loader />
+                              <Text $variation="600" $size="md">
+                                {(() => {
+                                  const toolInvocation = message.parts?.find(
+                                    (part) =>
+                                      part.type === 'tool-invocation' &&
+                                      part.toolInvocation.toolName !==
+                                        'document_parsing',
+                                  );
+                                  if (
+                                    toolInvocation?.type ===
+                                      'tool-invocation' &&
+                                    toolInvocation.toolInvocation.toolName ===
+                                      'summarize'
+                                  ) {
+                                    return t('Summarizing...');
+                                  }
+                                  return t('Search...');
+                                })()}
+                              </Text>
+                            </Box>
+                          )}
+                        {message.parts
+                          ?.filter(
+                            (part) =>
+                              part.type === 'reasoning' ||
+                              part.type === 'tool-invocation',
+                          )
+                          .map(
+                            (
+                              part: ReasoningUIPart | ToolInvocationUIPart,
+                              partIndex: number,
+                            ) =>
+                              part.type === 'reasoning' ? (
+                                <Box
+                                  key={`reasoning-${partIndex}`}
+                                  $background="var(--c--theme--colors--greyscale-100)"
+                                  $color="var(--c--theme--colors--greyscale-500)"
+                                  $padding={{ all: 'sm' }}
+                                  $radius="md"
+                                  $css="font-size: 0.9em;"
+                                >
+                                  {part.reasoning}
+                                </Box>
+                              ) : part.type === 'tool-invocation' &&
+                                isCurrentlyStreaming &&
+                                isLastAssistantMessageInConversation ? (
+                                <ToolInvocationItem
+                                  key={`tool-invocation-${partIndex}`}
+                                  toolInvocation={part.toolInvocation}
+                                  status={status}
+                                  hideSearchLoader={true}
+                                />
+                              ) : null,
+                          )}
+                      </Box>
+                      {message.role === 'assistant' &&
+                        !(
+                          isLastAssistantMessageInConversation &&
+                          status === 'streaming'
+                        ) && (
+                          <Box
+                            $css="color: #222631; font-size: 12px;"
+                            $direction="row"
+                            $align="center"
+                            $justify="space-between"
+                            $gap="6px"
+                            $margin={{ top: 'base' }}
+                          >
+                            <Box $direction="row" $gap="4px">
+                              <Box
+                                $direction="row"
+                                $align="center"
+                                $gap="4px"
+                                className="c__button--neutral action-chat-button"
+                                onClick={() => copyToClipboard(message.content)}
+                                onKeyDown={(e) => {
+                                  if (e.key === 'Enter' || e.key === ' ') {
+                                    e.preventDefault();
+                                    copyToClipboard(message.content);
+                                  }
+                                }}
+                                role="button"
+                                tabIndex={0}
+                              >
+                                <Icon
+                                  iconName="content_copy"
+                                  $theme="greyscale"
+                                  $variation="550"
+                                  $size="16px"
+                                  className="action-chat-button-icon"
+                                />
+                                {!isMobile && (
+                                  <Text $theme="greyscale" $variation="550">
+                                    {t('Copy')}
+                                  </Text>
+                                )}
+                              </Box>
+                              {message.parts?.some(
+                                (part) => part.type === 'source',
+                              ) &&
+                                (() => {
+                                  const sourceCount =
+                                    message.parts?.filter(
+                                      (part) => part.type === 'source',
+                                    ).length || 0;
+                                  return (
+                                    <Box
+                                      $direction="row"
+                                      $align="center"
+                                      $gap="4px"
+                                      className={`c__button--neutral action-chat-button ${isSourceOpen === message.id ? 'action-chat-button--open' : ''}`}
+                                      onClick={() => openSources(message.id)}
+                                      onKeyDown={(e) => {
+                                        if (
+                                          e.key === 'Enter' ||
+                                          e.key === ' '
+                                        ) {
+                                          e.preventDefault();
+                                          openSources(message.id);
+                                        }
+                                      }}
+                                      role="button"
+                                      tabIndex={0}
+                                    >
+                                      <Icon
+                                        iconName="book"
+                                        $theme="greyscale"
+                                        $variation="550"
+                                        $size="16px"
+                                        className="action-chat-button-icon"
+                                      />
+                                      <Text
+                                        $theme="greyscale"
+                                        $variation="550"
+                                        $weight="500"
+                                        $size="12px"
+                                      >
+                                        {t('Show')} {sourceCount}{' '}
+                                        {sourceCount !== 1
+                                          ? t('sources')
+                                          : t('source')}
+                                      </Text>
+                                    </Box>
+                                  );
+                                })()}
+                            </Box>
+                            <Box $direction="row" $gap="4px">
+                              {/* We should display the button, but disabled if no trace linked */}
+                              {conversationId &&
+                                message.id &&
+                                message.id.startsWith('trace-') && (
+                                  <FeedbackButtons
+                                    conversationId={conversationId}
+                                    messageId={message.id}
+                                  />
+                                )}
+                            </Box>
+                          </Box>
+                        )}
+                      {message.parts &&
+                        isSourceOpen === message.id &&
+                        (() => {
+                          const sourceParts = message.parts.filter(
+                            (part): part is SourceUIPart =>
+                              part.type === 'source',
+                          );
+                          return (
+                            <Box
+                              $css={`
+                              animation: fade-in 0.2s ease-out;
+                            `}
+                            >
+                              <SourceItemList
+                                parts={sourceParts}
+                                getMetadata={getMetadata}
+                              />
+                            </Box>
+                          );
+                        })()}
+                    </Box>
+                  </Box>
+                </Box>
               );
             })}
           </Box>
