+++ src/frontend/apps/conversations/src/features/chat/hooks/index.ts
@@ -1,3 +1,2 @@
 export { useChatScroll } from './useChatScroll';
 export { useSourceMetadataCache } from './useSourceMetadata';
-export { useModelSelection } from './useModelSelection';
